+++ src/ResourceManager/RecoveryServices/Commands.RecoveryServices.Backup.Test/ScenarioTests/IaasVm/Common.ps1
@@ -100,23 +100,35 @@ function Cleanup-ResourceGroup(
 		$vaults = Get-AzureRmRecoveryServicesVault -ResourceGroupName $resourceGroupName
 		foreach ($vault in $vaults)
 		{
-			Set-AzureRmRecoveryServicesVaultContext -Vault $vault
-			$containers = Get-AzureRmRecoveryServicesBackupContainer -ContainerType AzureVM
-			foreach ($container in $containers)
-			{
-				$items = Get-AzureRmRecoveryServicesBackupItem -Container $container -WorkloadType AzureVM
-				foreach ($item in $items)
-				{
-					Disable-AzureRmRecoveryServicesBackupProtection -Item $item -RemoveRecoveryPoints -Force
-				}
-			}
-
-			Remove-AzureRmRecoveryServicesVault -Vault $vault
+			Delete-Vault $vault
 		}
 	
 		# Cleanup RG. This cleans up all VMs and Storage Accounts.
 		Remove-AzureRmResourceGroup -Name $resourceGroupName -Force
 	}
+}
+
+function Delete-Vault($vault)
+{
+	$containers = Get-AzureRmRecoveryServicesBackupContainer `
+		-VaultId $vault.ID `
+		-ContainerType AzureVM
+	foreach ($container in $containers)
+	{
+		$items = Get-AzureRmRecoveryServicesBackupItem `
+			-VaultId $vault.ID `
+			-Container $container `
+			-WorkloadType AzureVM
+		foreach ($item in $items)
+		{
+			Disable-AzureRmRecoveryServicesBackupProtection `
+				-VaultId $vault.ID `
+				-Item $item `
+				-RemoveRecoveryPoints -Force
+		}
+	}
+
+	Remove-AzureRmRecoveryServicesVault -Vault $vault
 }
 
 function Create-VM(
@@ -202,25 +214,31 @@ function Enable-Protection(
 	$vault, 
 	$vm)
 {
-	Set-AzureRmRecoveryServicesVaultContext -Vault $vault | Out-Null
-
 	$container = Get-AzureRmRecoveryServicesBackupContainer `
+		-VaultId $vault.ID `
 		-ContainerType AzureVM `
 		-Name $vm.Name;
 
 	if ($container -eq $null)
 	{
-		$policy = Get-AzureRmRecoveryServicesBackupProtectionPolicy -Name "DefaultPolicy";
+		$policy = Get-AzureRmRecoveryServicesBackupProtectionPolicy `
+			-VaultId $vault.ID `
+			-Name "DefaultPolicy";
 	
 		Enable-AzureRmRecoveryServicesBackupProtection `
-			-Policy $policy -Name $vm.Name -ResourceGroupName $vm.ResourceGroupName | Out-Null
+			-VaultId $vault.ID `
+			-Policy $policy `
+			-Name $vm.Name `
+			-ResourceGroupName $vm.ResourceGroupName | Out-Null
 
 		$container = Get-AzureRmRecoveryServicesBackupContainer `
+			-VaultId $vault.ID `
 			-ContainerType AzureVM `
 			-Name $vm.Name;
 	}
 	
 	$item = Get-AzureRmRecoveryServicesBackupItem `
+		-VaultId $vault.ID `
 		-Container $container `
 		-WorkloadType AzureVM `
 		-Name $vm.Name
@@ -232,9 +250,9 @@ function Backup-Item(
 	$vault,
 	$item)
 {
-	Set-AzureRmRecoveryServicesVaultContext -Vault $vault | Out-Null
-
-	return Backup-AzureRmRecoveryServicesBackupItem -Item $item | Wait-AzureRmRecoveryServicesBackupJob
+	return Backup-AzureRmRecoveryServicesBackupItem `
+		-VaultId $vault.ID `
+		-Item $item | Wait-AzureRmRecoveryServicesBackupJob -VaultId $vault.ID
 }
 
 function Get-RecoveryPoint(
@@ -242,12 +260,14 @@ function Get-RecoveryPoint(
 	$item,
 	$backupJob)
 {
-	Set-AzureRmRecoveryServicesVaultContext -Vault $vault | Out-Null
-
 	$backupStartTime = $backupJob.StartTime.AddMinutes(-1);
 	$backupEndTime = $backupJob.EndTime.AddMinutes(1);
 
-	$rps = Get-AzureRmRecoveryServicesBackupRecoveryPoint -Item $item -StartDate $backupStartTime -EndDate $backupEndTime
+	$rps = Get-AzureRmRecoveryServicesBackupRecoveryPoint `
+		-VaultId $vault.ID `
+		-Item $item `
+		-StartDate $backupStartTime `
+		-EndDate $backupEndTime
 	
 	return $rps[0]
 }