+++ src/ResourceManager/RecoveryServices/Commands.RecoveryServices.Backup/Cmdlets/Container/UnregisterAzureRmRecoveryServicesBackupManagementServer.cs
@@ -16,6 +16,7 @@ using System;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models;
 using Microsoft.Azure.Commands.RecoveryServices.Backup.Properties;
+using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
 
 namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Container
 {
@@ -25,7 +26,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Container
     [Cmdlet(VerbsLifecycle.Unregister, "AzureRmRecoveryServicesBackupManagementServer",
         SupportsShouldProcess = true)]
     public class UnregisterAzureRmRecoveryServicesBackupManagementServer
-        : RecoveryServicesBackupCmdletBase
+        : RSBackupVaultCmdletBase
     {
         /// <summary>
         /// The backup management server to be unregistered from the vault.
@@ -47,6 +48,10 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Container
             {
                 base.ExecuteCmdlet();
 
+                ResourceIdentifier resourceIdentifier = new ResourceIdentifier(VaultId);
+                string vaultName = resourceIdentifier.ResourceName;
+                string resourceGroupName = resourceIdentifier.ResourceGroupName;
+
                 if ((AzureRmBackupManagementServer.BackupEngineType !=
                     BackupEngineType.DpmBackupEngine.ToString() &&
                     AzureRmBackupManagementServer.BackupEngineType !=
@@ -63,7 +68,10 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Container
                 }
 
                 string azureRmBackupManagementServer = AzureRmBackupManagementServer.Name;
-                ServiceClientAdapter.UnregisterContainers(azureRmBackupManagementServer);
+                ServiceClientAdapter.UnregisterContainers(
+                    azureRmBackupManagementServer,
+                    vaultName: vaultName,
+                    resourceGroupName: resourceGroupName);
 
                 if (PassThru.IsPresent)
                 {