+++ src/ResourceManager/RecoveryServices/Commands.RecoveryServices.Backup/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Added -Vault parameter to RecoveryServices.Backup cmdlets. When passed, this will override the Set-AzureRmRecoveryServicesContext cmdlet.
 
 ## Version 4.1.3
 * Set minimum dependency of module to PowerShell 5.0